src/main/java/com/ppt/qa/pages/LoginPage.java
package com.ppt.qa.pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.ppt.qa.base.TestBase;
//import com.ppt.qa.pages.HomePage;

public class LoginPage extends TestBase {
	
	//Page Factory
	@FindBy(xpath="//input[@name='username']")
	WebElement username;
	
	@FindBy(xpath="//input[@name='password']")
	WebElement password;
	
	@FindBy(xpath="//input[@type='submit']")
	WebElement loginBtn;
	
	@FindBy(xpath="//a[contains(text(),'Sign')]")
	WebElement signUpLink;
	
	@FindBy(xpath="//a[@class='navbar-brand']")
	WebElement crmlogoImage;
	
	//Initializing the Page Objects
	public LoginPage()
	{
		PageFactory.initElements(driver, this);
	}
	
	//Actions
	public String validateLoginPageTitle()
	{
		return driver.getTitle();
	}
	
	public boolean validateCRMImg()
	{
		return crmlogoImage.isDisplayed();
	}
//	public HomePage login(String un, String pwd)
//	{
//		username.sendKeys(un);
//		password.sendKeys(pwd);
//		loginBtn.click();
//		return new HomePage();
//	}
	
}
